Signup page now detects duplicate students

+++ sportsday_site/views.py
@@ -42,15 +42,30 @@ def student_signups(request,houseid,category,gender):
     valid_students=Student.objects.filter(gender=gender,house=houseid,category=category)
     gender_string="Boys" if gender=="M" else "Girls" if gender=="F" else "Mixed"
     if request.method=="POST":
+        valid_input=True
+        new_signups={}
         for event in valid_events:
-            curev_signups=Signup.objects.filter(signed_event=event,signed_event__gender=gender,signed_student__house=houseid)
-            curev_signups.delete()#replace with new signups
+            student_set=set()
             for slotnum in range(2):
                 form_field=f"event_{event.pk}_slot_{slotnum+1}"
                 if form_field in request.POST:
-                    student_obj=Student.objects.get(pk=request.POST[form_field])
+                    result=request.POST[form_field]
+                    if result in student_set:
+                        valid_input=False
+                        break
+                    student_set.add(request.POST[form_field])
+            if not valid_input:
+                break
+            new_signups[event]=student_set
+        if valid_input:
+            for event in valid_events:
+                curev_signups=Signup.objects.filter(signed_event=event,signed_event__gender=gender,signed_student__house=houseid)
+                curev_signups.delete()#replace with new signups
+                for student_pk in student_set:
+                    student_obj=Student.objects.get(pk=student_pk)
                     Signup.objects.create(signed_student=student_obj,signed_event=event)
-
+        else:
+            return HttpResponse("error")
         print("form submitted")
 
         return HttpResponseRedirect(request.path_info)
